Modified PdfName to store it's value in a std::string as well as supporting construction and equality from a std::string Added support to PdfOutputDevice to write to an arbitrary std::ostream Removed direct std::string support from PdfOutputDevice (since you should now use a std::ostringstream) Removed more compiler warnings Cleaned up misc. things in the code
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@33 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfWriter.h
@@ -235,7 +235,7 @@ class PdfWriter {
     PdfError WritePdfTableOfContents( PdfOutputDevice* pDevice );
 
  protected:
-    unsigned int    m_nObjectCount;
+    size_t			m_nObjectCount;
     TVecObjects     m_vecObjects;
 
  private:
